fix test failure in pre commit
